Add new aggressive flow when encountering police
Instead of just asking you to follow with options to escape or follow,
add more aggressive encounter.

Cop tells player to put their hands in the air, at which point the
player can comply or try to escape and fight like normal

After putting hands in air, Police tell player to get on their knees.
This is the player's last chance to try to fight and escape.

After getting on knees, player is told to put their hands behind their
back, at which point they will be handcuffed. If they try to escape once
on their knees, the police will stop them and punish them with manacles.

Once the aggressive encounter ends in defeat, either through submission
or losing fight, the player is taken to the prison where the normal
event flow is followed.

If the player enters the fight, normal event flow is followed.